Documentation Update
This update to documentation was done to prepare for verification with Discord.
Added a mention of the Privacy Policy and Terms in the README.
Small update to the privacy policy, fixing a placeholder text I forgot about and detailing why data may be kept in accordance with GDPR.
Added a Terms of Service/User Agreement.